feat(publication): add new components for different versions of PublicationPage
Two new components, PublicationPage and PublicationPageWrapper, have been added to the publication app. These components are designed to render different versions of the PublicationPage based on the version specified in the component path. This addition enhances the flexibility and modularity of the publication app by allowing for easy switching between different versions of the PublicationPage component.