chore(dicty-frontpage): update coverage-report script to use npx nyc@17 for generating coverage report in lcov format
The coverage-report script in the package.json file has been updated to use npx nyc@17 instead of yarn c8 for generating the coverage report in the lcov format. This change was made to ensure compatibility with the latest version of the nyc package and to improve the consistency of the project's tooling.