feat(editor-toolbar): add AttachFileButton component for inserting images in the editor toolbar
The new AttachFileButton component allows users to easily insert images into the editor toolbar. It includes a button with an "Image" label and an icon that opens an ImageUploadDialog when clicked. This feature enhances the functionality of the editor toolbar by providing a convenient way to add images to the content.