feat(callback.tsx): add Callback component to handle sign-in callback and display loading screen
The Callback component is added to the genome-page application to handle the callback after a successful sign-in. It utilizes the useHandleSignInCallback hook from the @logto/react library to manage the sign-in process and redirects the user to the home page ("/") upon completion. The component renders a LoadingDisplay from the @dictybase/ui-common package to show a loading screen while the sign-in process is ongoing.

Resolves #680