feat(Frontpage): create image slideshow component
This commit introduces a new Slideshow component that displays a carousel of curated images related to dictyBase. The component utilizes the react-responsive-carousel library for the carousel functionality and Material-UI for styling. The images, titles, and descriptions are stored in an array of objects, which are then mapped over to render each slide in the carousel. The carousel has autoplay enabled with an interval of 5 seconds and loops infinitely. The component is designed to enhance the visual appeal of the frontpage and provide informative content about dictyBase.