feat(DictyInfoWithAuth.tsx): create DictyInfoWithAuth component
This component conditionally renders either the DictyInfo or
AuthorizedDictyInfo component based on user authorization. It uses the
useAuthorization hook to check if the user has the "content-admin" role.
If the user is authorized, it renders the AuthorizedDictyInfo component,
otherwise it renders the DictyInfo component. Both components receive
the result of the useContentBySlugQuery hook, which fetches content
data.