feat(homeStyles.ts): add icon style to useStyles to control icon size
The new icon style allows the icon size to be controlled by the h2 font size in the theme. If the h2 font size is not defined, the icon size defaults to 24px. This change provides more flexibility in styling the icons. The fp-ts library is used to handle the optional value of the h2 font size.