feat(StockCenterInfoWithAuth.tsx): create StockCenterInfoWithAuth component
This component conditionally renders either the StockCenterInfo or the AuthorizedStockCenterInfo component based on user authorization. It uses the useAuthorization hook to check if the user has the "content-admin" role. If the user is authorized, it renders the AuthorizedStockCenterInfo component; otherwise, it renders the StockCenterInfo component. This allows for different content to be displayed based on user roles.