refactor(FontSizeDropdown.tsx): change font size array generation to return objects with value and label
The font size array generation was refactored to return an array of objects, each with a `value` and a `label` property. This improves the readability and maintainability of the code, as it makes it clearer what each value in the array represents. It also makes it easier to use the array in the `MenuItem` components, as the `value` and `label` properties can be directly accessed.