feat(editor-toolbar): add download attribute to DictyLinkNode
This commit introduces the ability to add a download attribute to
DictyLinkNode. This allows users to specify a filename for the linked
resource when it is downloaded.

The following changes were made:

- Added a `__download` property to the `DictyLinkNode` class to store
  the download filename.
- Added `setDownload` and `getDownload` methods to the
  `DictyLinkNode` class to set and retrieve the download filename.
- Updated the `clone` method to copy the `__download` property when
  cloning a `DictyLinkNode`.
- Updated the `createDOM` method to set the `download` attribute of
  the anchor element when the `__download` property is set.
- Updated the `updateDOM` method to update the `download` attribute of
  the anchor element when the `__download` property changes.

These changes allow users to create links that automatically trigger a
download with a specified filename when clicked.