feat(DictyLinkNode): add download attribute to DictyLinkNode
This commit adds a download attribute to the DictyLinkNode,
allowing users to specify a file to be downloaded when the link
is clicked. It also includes the necessary serialization and
deserialization logic to persist the download attribute.
The download attribute is now included when exporting the node
to JSON and is used when importing the node from JSON.