feat(editor-toolbar): enhance DictyLinkNode and FileUploadDialog
This commit enhances the `DictyLinkNode` component by adding a newline for better readability. Additionally, it improves the `FileUploadDialog` component by using `pipe` and `OgetOrElse` to handle the `initialFileName` prop, ensuring a default value is provided when `fileName` is not available. This prevents potential errors and improves the robustness of the component.