feat(editor): replace LinkNode with DownloadLinkNode in editor config
This commit introduces a new `DownloadLinkNode` that extends the `LinkNode` to support the `download` attribute. This allows users to specify a filename for the downloaded file when the link is clicked. The `LinkNode` is replaced with the `DownloadLinkNode` in the editor configuration, so all links created in the editor will now be `DownloadLinkNode`s.