feat(FileUploadDialog.tsx): add upload as filename functionality
Adds a new feature to the FileUploadDialog component that allows
users to specify a custom filename for the uploaded file. This
filename is stored in the `uploadAsFileName` state variable and
passed to the `createFileUploadFunction` function. The
`Upload` component is also updated to include a new prop for
setting the `uploadAsFileName`.

This change allows users to have more control over the naming of
their uploaded files.