feat(editor-toolbar): add file name input to upload dialog
Adds a new text field to the upload dialog that allows the user to
specify the name of the file that will be downloaded. The file name
input is only visible when a file has been selected.

The file name is stored in the `uploadAsName` state variable, which is
updated when the user types in the text field.

This change allows the user to control the name of the file that is
downloaded, which can be useful for providing a more descriptive or
user-friendly file name.