feat(editor-toolbar): create SelectAndUpload component
This component allows users to select a file from their
computer and upload it to the server. It uses the
`FileSelect` component to display the file selection
input and the `Upload` component to handle the file
upload. The component also uses the
`useValidateUploadName` hook to validate the file name
before uploading it.