chore: update dependencies in yarn.lock
This commit updates various dependencies in the `yarn.lock` file.
The updates include both minor and patch version bumps for packages
such as `@eslint/*`, `browserslist`, `core-js-compat`,
`eslint-plugin-unicorn`, and many others.

The dependency updates are performed to ensure that the project is
using the latest versions of its dependencies, which may include bug
fixes, performance improvements, and new features. Updating
dependencies also helps to maintain compatibility with other
packages and tools in the ecosystem.