feat(useAuthorizedUpdateContent.ts): return data from update
The hook now returns the data from the update instead of the entire
update object. This simplifies the usage of the hook and reduces the
amount of data that needs to be processed.
Also, a new type `UpdateContentError` is added to provide a more
specific type for update content errors.