refactor(hook): move FieldOption component from ui-dsc to hook package
The FieldOption component is a visual component that is used in the
Searchbox component. It is used for displaying the field options in the
Autocomplete component as well as the `active chip value` for the
Autocomplete's tags. This component was previously located in the
`ui-dsc` package, but it has been moved to the `hook` package because
it is only used by the `useSearchWithRouter` hook. A unit test was
added for the FieldOption component to ensure that it renders the text
provided to the label prop. The import path for the FieldOption
component was updated in the `useSearchWithRouter` hook. The
FieldOption component was removed from the `ui-dsc` package.