feat(Edit.tsx, Show.tsx): set fetchPolicy to cache-and-network
This change modifies the fetch policy for the useContentBySlugQuery hook in both the Edit and Show components. By setting the fetchPolicy to cache-and-network, the components will now first attempt to retrieve data from the cache. If the data is not available in the cache or is stale, the hook will then fetch the data from the network. This ensures that the components always have the most up-to-date data while also improving performance by reducing the number of network requests.