feat(EditView.tsx): add autosave states to EditActionBar
This commit adds autosave states to the EditActionBar component.
The autosave states are used to display the status of the autosave process.
The states are:
- waiting: when the autosave is waiting to be saved
- loading: when the autosave is being saved
- error: when the autosave has failed
- success: when the autosave has been saved

The autosave states are displayed using the PendingChanges, WaitingChanges, and ProgressSaved components from the @dictybase/ui-common package.
The ExitEditingButton component is also added to the EditActionBar component.
The UpdateButton component now accepts a canSave prop, which is set to false.
The Snackbar component has been removed.
The timeSince function has been replaced with the formatDistance function from the date-fns package.
The useConfirmNavigation hook has been removed.