feat: add SavingError component to ui-common package
This commit introduces a new `SavingError` component to the
`ui-common` package. This component displays an error icon,
providing a visual cue to the user that an error occurred during
the saving process. The component is also exported from the
`ui-common` package's index file, making it readily available
for use in other parts of the application.