feat(editor): add onChange handler to editor
This change adds an `onChange` handler to the editor, which
allows the parent component to be notified when the editor's
content changes. This is useful for implementing features such
as autosaving and collaboration.
The `OnChangePlugin` component is used to register the
`onChange` handler with the editor. The `ignoreSelectionChange`
property is set to `true` to prevent the handler from being
called when the selection changes.
The `handleChange` property is added to the `EditorProperties`
type to allow the parent component to pass in the `onChange`
handler.
The `id` property is added to the `ContentEditable` component
to allow the parent component to access the editor's content.